requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-cloud-storage==2.18.2
google-cloud-firestore==2.19.0
google-cloud-secret-manager==2.21.1
requests==2.32.3
pydantic==2.9.2
sse-starlette==2.1.3
